requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
asgiref==3.8.1
attrs==25.1.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.1
django-ckeditor==6.7.1
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
django-js-asset==2.2.0
django-modeltranslation==0.19.9
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
drf-spectacular==0.28.0
drf-yasg==1.21.8
frozenlist==1.5.0
geographiclib==2.0
geopy==2.4.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
multidict==6.1.0
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
propcache==0.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyngrok==7.2.3
python-decouple==3.8
python-dotenv==1.0.1
python-gettext==5.0
python-slugify==8.0.4
python3-openid==3.2.0
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
six==1.16.0
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.1
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.3.0
yarl==1.18.3
